refactors and new implentations

+++ lib/services/file_handler.dart
@@ -4,36 +4,18 @@ import 'package:flutter/foundation.dart';
 import 'package:hangr/services/calendar_map.dart';
 import 'package:hangr/services/outfit.dart';
 import 'package:hangr/services/wearable.dart';
-import 'package:path_provider/path_provider.dart';
 
 class FileHandler {
   static const _outfitsPath = 'outfits.json';
   static const _wearablesPath = 'wearables.json';
   static const _calendarMapPath = 'calendar_map.json';
-  static late final String _dirPath;
+  final String _directoryPath;
 
-  FileHandler() {
-    _getDirPath().then((path) => _dirPath = path).onError((error, stackTrace) {
-      return _dirPath = "";
-    });
-  }
+  FileHandler(this._directoryPath);
 
-  Future<String> _getDirPath() async {
-    String path = "";
-    try {
-      final localDir = await getApplicationDocumentsDirectory();
-      path = localDir.path;
-    } catch (e) {
-      if (kDebugMode) {
-        print(e);
-      }
-    }
-    return path;
-  }
-
-  Future<CalendarMap?> readCalendarMap() async {
-    final f = File('$_dirPath/$_calendarMapPath');
-    CalendarMap? map;
+  Future<CalendarMap> readCalendarMap() async {
+    final f = File('$_directoryPath/$_calendarMapPath');
+    CalendarMap map = CalendarMap.empty();
 
     if (await f.exists()) {
       final content = await f.readAsString();
@@ -42,16 +24,17 @@ class FileHandler {
     return map;
   }
 
-  Future<List<Outfit>?> readOutfits() async {
-    final File f = File('$_dirPath/$_outfitsPath');
-    List<Outfit>? outfits;
+  Future<List<Outfit>> readOutfits() async {
+    final File f = File('$_directoryPath/$_outfitsPath');
+    List<Outfit> outfits = [];
 
     try {
       if (await f.exists()) {
         await f.readAsString().then((contents) {
           final Iterable l = jsonDecode(contents) as Iterable<dynamic>;
           outfits = List<Outfit>.from(
-              l.map((model) => Outfit.fromJson(model as Map<String, dynamic>)),);
+            l.map((model) => Outfit.fromJson(model as Map<String, dynamic>)),
+          );
         });
       }
     } on Exception catch (e) {
@@ -62,16 +45,19 @@ class FileHandler {
     return outfits;
   }
 
-  Future<List<Wearable>?> readWearables() async {
-    final File f = File('$_dirPath/$_wearablesPath');
-    List<Wearable>? wearables;
+  Future<List<Wearable>> readWearables() async {
+    final File f = File('$_directoryPath/$_wearablesPath');
+    List<Wearable> wearables = [];
 
     try {
       if (await f.exists()) {
         await f.readAsString().then((contents) {
           final Iterable l = jsonDecode(contents) as Iterable<dynamic>;
-          wearables = List<Wearable>.from(l.map(
-              (model) => Wearable.fromJson(model as Map<String, dynamic>),),);
+          wearables = List<Wearable>.from(
+            l.map(
+              (model) => Wearable.fromJson(model as Map<String, dynamic>),
+            ),
+          );
         });
       }
     } on Exception catch (e) {
@@ -84,56 +70,59 @@ class FileHandler {
 
   Future<bool> writeWearables(List<Wearable> wearables) async {
     try {
-      File f = File('$_dirPath/$_wearablesPath');
-      f = await f.writeAsString(json.encode(wearables));
+      await File('$_directoryPath/$_wearablesPath')
+          .writeAsString(json.encode(wearables));
+      return await File('$_directoryPath/$_wearablesPath').exists();
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
       return false;
     }
-    return true;
   }
 
   Future<bool> writeOutfits(List<Outfit> outfits) async {
     try {
-      File f = File('$_dirPath/$_outfitsPath');
-      f = await f.writeAsString(json.encode(outfits));
+      await File('$_directoryPath/$_outfitsPath')
+          .writeAsString(json.encode(outfits));
+      return await File('$_directoryPath/$_outfitsPath').exists();
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
       return false;
     }
-    return true;
   }
 
   Future<bool> writeCalendarMap(CalendarMap map) async {
     try {
-      File f = File('$_dirPath/$_calendarMapPath');
-      f = await f.writeAsString(json.encode(map));
+      await File('$_directoryPath/$_calendarMapPath')
+          .writeAsString(json.encode(map));
+      return await File('$_directoryPath/$_calendarMapPath').exists();
     } on Exception catch (e) {
       if (kDebugMode) {
         print(e);
       }
       return false;
     }
-    return true;
   }
 
   Future<bool> deleteAllFiles() async {
     try {
-      final File wearables = File('$_dirPath/$_wearablesPath');
-      final File outfits = File('$_dirPath/$_outfitsPath');
-      final File map = File('$_dirPath/$_calendarMapPath');
+      final File wearables = File('$_directoryPath/$_wearablesPath');
+      final File outfits = File('$_directoryPath/$_outfitsPath');
+      final File map = File('$_directoryPath/$_calendarMapPath');
 
-      wearables.delete();
-      outfits.delete();
-      map.delete();
-    } on Exception catch (e) {
-      if (kDebugMode) {
-        print(e);
+      if (await wearables.exists()) {
+        wearables.delete();
       }
+      if (await outfits.exists()) {
+        outfits.delete();
+      }
+      if (await map.exists()) {
+        map.delete();
+      }
+    } on Exception {
       return false;
     }
     return true;
